+++ src/components/organisms/Schedule.tsx
@@ -1,10 +1,6 @@
+import eventDetails from "@/data/eventDetails.json";
 import { mq } from "@/themes/settings/breakpoints";
-import { brand, neutral, themeLight } from "@/themes/settings/color";
 import { css } from "@emotion/react";
-import Image from "next/image";
-import Link from "next/link";
-import seikaiha from "public/images/seikaiha.png";
-import { GoInfo } from "react-icons/go";
 import { HiCalendarDays, HiOutlineMapPin } from "react-icons/hi2";
 import { Socials } from "./Socials";
 
@@ -16,7 +12,7 @@ const Info = () => {
         display:flex;
         flex-direction:column;
         justify-content: center;
-        padding: 4rem;
+        padding: 4rem 2rem 2rem 0;
         text-align: center;
 
         ${mq.laptop} {
@@ -25,36 +21,70 @@ const Info = () => {
         }
       `}
       >
-        <div css={css`align-self:center;`}>
+        <div css={css`align-self:start;`}>
           <HiCalendarDays size={60} />
         </div>
-        <div css={css`padding: 0 1rem;`}>
-          <h2 css={css`margin:0;`}>Hackathon</h2>
-          <p css={css`font-size: 1rem; margin:0;`}>23-25 August 2024</p>
-        </div>
-        <div css={css`padding: 0 1rem;`}>
-          <h2 css={css`margin:0;`}>Conference</h2>
-          <p css={css`font-size: 1rem; margin:0;`}>26 August 2024</p>
+
+        <div
+          css={css`
+          display: flex;
+          flex-direction: column;
+          padding: 0 1rem;
+        `}
+        >
+          <div>
+            <h2 css={css`margin:0;`}>Conference</h2>
+            <p css={css`font-size: 1rem; margin:0;`}>
+              {eventDetails.eventDate.conference}
+            </p>
+          </div>
+          <div css={css`margin-top: 1rem;`}>
+            <h2 css={css`margin:0;`}>Hackathon</h2>
+            <p css={css`font-size: 1rem; margin:0;`}>
+              {eventDetails.eventDate.hackathon}
+            </p>
+          </div>
         </div>
         <div css={css`height:2rem; ${mq.laptop}{width: 2rem;}`} />
-        <div css={css`align-self:center;`}>
+        <div css={css`align-self:start;`}>
           <HiOutlineMapPin size={60} />
         </div>
-        <div css={css`padding: 0 1rem;`}>
-          <h2 css={css`margin:0;`}>Digital Garage</h2>
-          <a
-            css={css`text-decoration: none; :hover {text-decoration: underline;} ;`}
-            href="https://maps.app.goo.gl/aCRdR2pX4qRaUz6MA"
-            target="_blank"
-            rel="noreferrer"
-          >
-            <p css={css`font-size: 1rem; margin:0;`}>
-              15-1 Udagawacho, Shibuya City, Tokyo
-            </p>
-          </a>
+        <div
+          css={css`
+          display: flex;
+          flex-direction: column;
+          padding: 0 1rem;
+        `}
+        >
+          <div css={css``}>
+            <h2 css={css`margin:0;`}>United Nations University</h2>
+            <a
+              css={css`text-decoration: none; :hover {text-decoration: underline;} ;`}
+              href="https://maps.app.goo.gl/uZKbVZm5wmGyCsAg6"
+              target="_blank"
+              rel="noreferrer"
+            >
+              <p css={css`font-size: 1rem; margin:0;`}>
+                5-53-70 Jingumae, Shibuya, Tokyo
+              </p>
+            </a>
+          </div>
+          <div css={css`margin-top: 1rem;`}>
+            <h2 css={css`margin:0;`}>Digital Garage</h2>
+            <a
+              css={css`text-decoration: none; :hover {text-decoration: underline;} ;`}
+              href="https://maps.app.goo.gl/aCRdR2pX4qRaUz6MA"
+              target="_blank"
+              rel="noreferrer"
+            >
+              <p css={css`font-size: 1rem; margin:0;`}>
+                15-1 Udagawacho, Shibuya, Tokyo
+              </p>
+            </a>
+          </div>
         </div>
         <div css={css`height:2rem; ${mq.laptop}{width: 2rem;}`} />
-        <div css={css`align-self:center;`}>
+        {/* <div css={css`align-self:center;`}>
           <GoInfo size={60} />
         </div>
         <div css={css`padding: 0 1rem;`}>
@@ -70,7 +100,7 @@ const Info = () => {
               Click to see all events
             </a>
           </p>
-        </div>
+        </div> */}
       </div>
     </div>
   );
@@ -81,12 +111,6 @@ const ScheduleSection = () => {
     <section
       id="schedule"
       css={css`
-        background-color: ${brand.Miyabi};
-        background-image: url(${seikaiha.src});
-        background-position: center;
-        background-repeat: repeat;
-        background-size: cover;
-        padding: 2rem 0;
         text-align: center;
         width: 100%;
       `}
