+++ src/components/organisms/Statement.tsx
@@ -1,19 +1,34 @@
+import eventDetails from "@/data/eventDetails.json";
 import { mq } from "@/themes/settings/breakpoints";
 import { brand } from "@/themes/settings/color";
 import { css } from "@emotion/react";
 import Image from "next/image";
 import ETHDiamondLogo from "public/logo/ETHDiamondLogo.png";
+import ETHTokyoLogo from "public/logo/ETHTokyoLogo.png";
 
 const StatementSection = () => {
+  const EventLogo = () => {
+    return (
+      <Image
+        src={ETHTokyoLogo}
+        alt="ETHTokyo logo"
+        css={css`
+          height: auto;
+          max-width: 256px;
+          padding-left: 11px;
+          padding-top: 1rem;
+        `}
+      />
+    );
+  };
+
   return (
     //background: linear-gradient(45deg, , );
     <section
       id="statement"
       css={css`
-        background: linear-gradient(0deg, #FF5544, #FF6F9D);
+        background: linear-gradient(0deg, ${brand.Shuiro}, #000);
         border-radius: 0 0 0 0 / 0 0 100% 100%;
-        clip-path:  ellipse(120% 100% at 50% 100%);
-        margin-top: -4rem;
         z-index: 10;
       `}
     >
@@ -28,10 +43,11 @@ const StatementSection = () => {
         }
       `}
       >
-        <Image src={ETHDiamondLogo} alt="Ethereum diamond" />
-        <h2 css={css`color: ${brand.Miyabi}; font-weight: 300;`}>
-          Welcome and join us!
-        </h2>
+        {/* <Image src={ETHDiamondLogo} alt="Ethereum diamond" /> */}
+        <div css={css`padding: 2rem 0;`}>
+          <EventLogo />
+        </div>
+        <h2 css={css`font-weight: 600;`}>Welcome and join us!</h2>
         <p css={css`text-align: justify; ${mq.laptop}{font-size: 1.25rem;}`}>
           ETHTokyo is a engaging hackathon for the global Ethereum community
           where people with all sorts of backgrounds, ideas, and skills come
@@ -41,6 +57,7 @@ const StatementSection = () => {
           foster the creation of amazing new innovations for the future. Join us
           and be a part of shaping what's next in the Ethereum universe!
         </p>
+        Further details will be announced soon.
       </div>
     </section>
   );
