+++ src/data/eventDetails.json
@@ -1,0 +1,8 @@
+{
+  "eventDate": {
+    "duration": "Sep 12-15, 2025",
+    "conference": "September 12",
+    "hackathon": "September 13-15"
+  },
+  "eventLocation": "Tokyo, Japan"
+}
